+++ packages/core/kind/kind.js
@@ -9,6 +9,7 @@ import {createContext, use, Component as ReactComponent} from 'react';
 
 import useHandlers from '../useHandlers';
 import Handlers from '../useHandlers/Handlers';
+import {checkPropTypes} from '../util';
 
 import computed from './computed';
 import styles from './styles';
@@ -163,16 +164,32 @@ const kind = (config) => {
 				...boundHandlers
 			};
 
+			if (defaultProps) {
+				Object.keys(defaultProps).forEach(key => {
+					// eslint-disable-next-line no-undefined
+					if (merged[key] === undefined) {
+						merged[key] = defaultProps[key];
+					}
+				});
+			}
+
+			checkPropTypes(Component, merged);
+
 			return renderKind(merged, ctx);
 		};
 	} else {
 		Component = class extends ReactComponent {
 			static contextType = contextType;
 
-			constructor () {
-				super();
+			constructor (props) {
+				super(props);
+				checkPropTypes(this, props);
 
 				this.handlers = new Handlers(handlers);
+			}
+
+			componentDidUpdate (prevProps) {
+				checkPropTypes(this, this.props, prevProps);
 			}
 
 			render () {