+++ packages/i18n/tests/ui/config/wdio.docker.conf.js
@@ -1,3 +1,0 @@
-import {uiDockerConfig} from '@enact/ui-test-utils/ui/wdio.docker.conf.js';
-
-export default uiDockerConfig;