+++ packages/ui/VirtualList/VirtualList.js
@@ -9,7 +9,7 @@
  * @exports VirtualListBasic
  */
 
-import {setDefaultProps} from '@enact/core/util';
+import {checkPropTypes, setDefaultProps} from '@enact/core/util';
 import PropTypes from 'prop-types';
 
 import {ResizeContext} from '../Resizable';
@@ -49,6 +49,7 @@ const virtualListDefaultProps = {
  * @public
  */
 const VirtualList = (props) => {
+	checkPropTypes(VirtualList, props);
 	// Hooks
 
 	const virtualListProps = setDefaultProps(props, virtualListDefaultProps);
@@ -327,6 +328,7 @@ const virtualGridListDefaultProps = {
  * @public
  */
 const VirtualGridList = (props) => {
+	checkPropTypes(VirtualList, props);
 	const virtualGridListProps = setDefaultProps(props, virtualGridListDefaultProps);
 
 	const {