+++ packages/ui/useScroll/Scrollbar.js
@@ -1,5 +1,5 @@
 import classNames from 'classnames';
-import {Job} from '@enact/core/util';
+import {checkPropTypes, Job} from '@enact/core/util';
 import PropTypes from 'prop-types';
 import {memo, useCallback, useEffect, useLayoutEffect, useRef} from 'react';
 
@@ -134,6 +134,7 @@ const Scrollbar = memo(({corner = false, css = componentCss, minThumbSize = 18, 
 		scrollbarProps,
 		scrollbarTrackProps
 	} = useScrollbar(props);
+	checkPropTypes(Scrollbar, props);
 
 	return (
 		<div {...restProps} {...scrollbarProps}>